refactor: split poly.utils module into a poly.utils package of focused modules
utils.py mixed six unrelated concerns in one 546-line module. Each now lives
in its own submodule: credentials, merge (3-way merge), stub_gen (_gen package
generation), decorators (func_* stubs + export_decorators, colocated because
export resolves sources via globals()), variable_references and commands.

The package __init__ re-exports all public names, so every existing
`poly.utils` import (internal and external) keeps working unchanged.
